db_init/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
azure-core==1.10.0
    # via azure-identity
azure-identity==1.5.0
    # via azure-kusto-data
azure-kusto-data==2.0.0
    # via -r requirements.in
certifi==2020.12.5
    # via requests
cffi==1.14.4
    # via cryptography
chardet==4.0.0
    # via requests
cryptography==3.3.1
    # via
    #   azure-identity
    #   msal
    #   pyjwt
idna==2.10
    # via requests
msal-extensions==0.3.0
    # via azure-identity
msal==1.8.0
    # via
    #   azure-identity
    #   azure-kusto-data
    #   msal-extensions
portalocker==1.7.1
    # via msal-extensions
pycparser==2.20
    # via cffi
pyjwt[crypto]==1.7.1
    # via msal
python-dateutil==2.8.1
    # via azure-kusto-data
requests==2.25.1
    # via
    #   azure-core
    #   azure-kusto-data
    #   msal
six==1.15.0
    # via
    #   azure-core
    #   azure-identity
    #   cryptography
    #   python-dateutil
    #   tenacity
tenacity==6.3.1
    # via -r requirements.in
urllib3==1.26.3
    # via requests
